Fixes #15489: Remove unused dependencies from relayd
