+++ relay/sources/relayd/Cargo.toml
@@ -26,6 +26,7 @@ nom = "5.0"
 flate2 = { version = "1.0", features = ["zlib"], default-features = false }
 diesel = { version = "1.4", features = ["postgres", "chrono", "r2d2"] }
 #### Bad (=Not widely used or <1.0)
+log = "0.4"
 structopt = { version = "0.2", features = [ "paw" ] }
 paw = "1.0"
 toml = "0.5"
@@ -38,10 +39,9 @@ tokio-threadpool = "0.1"
 tokio-process = "0.2"
 tokio-io = "0.1"
 # Compile dev and release with trace logs enabled
-log = "0.4"
 tracing = { version = "0.1", features = ["max_level_trace", "release_max_level_trace"] }
-tracing-fmt = { git = "https://github.com/tokio-rs/tracing" }
-tracing-log = { git = "https://github.com/tokio-rs/tracing" }
+tracing-fmt = "0.0.1-alpha.3"
+tracing-log = "0.0.1-alpha.1"
 futures = "0.1"
 warp = "0.1"
 reqwest = "0.9"
